% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{AccessConfig}
\alias{AccessConfig}
\title{Replica Pool API Objects 
The Replica Pool API allows users to declaratively provision and manage groups of Google Compute Engine instances based on a common template.}
\usage{
AccessConfig(name = NULL, natIp = NULL, type = NULL)
}
\arguments{
\item{name}{Name of this access configuration}

\item{natIp}{An external IP address associated with this instance}

\item{type}{Type of this access configuration file}
}
\value{
AccessConfig object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:08:46
filename: /Users/mark/dev/R/autoGoogleAPI/googlereplicapoolv1beta1.auto/R/replicapool_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AccessConfig Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Compute Engine network accessConfig. Identical to the accessConfig on corresponding Compute Engine resource.
}
