% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{Action}
\alias{Action}
\title{Action Object}
\usage{
Action(commands = NULL, envVariables = NULL, timeoutMilliSeconds = NULL)
}
\arguments{
\item{commands}{A list of commands to run, one per line}

\item{envVariables}{A list of environment variables to use for the commands in this action}

\item{timeoutMilliSeconds}{If an action's commands on a particular replica do not finish in the specified timeoutMilliSeconds, the replica is considered to be in a FAILING state}
}
\value{
Action object
}
\description{
Action Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An action that gets executed during initialization of the replicas.
}
