% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{DiskAttachment}
\alias{DiskAttachment}
\title{DiskAttachment Object}
\usage{
DiskAttachment(deviceName = NULL, index = NULL)
}
\arguments{
\item{deviceName}{The device name of this disk}

\item{index}{A zero-based index to assign to this disk, where 0 is reserved for the boot disk}
}
\value{
DiskAttachment object
}
\description{
DiskAttachment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies how to attach a disk to a Replica.
}
