% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{EnvVariable}
\alias{EnvVariable}
\title{EnvVariable Object}
\usage{
EnvVariable(hidden = NULL, name = NULL, value = NULL)
}
\arguments{
\item{hidden}{Deprecated, do not use}

\item{name}{The name of the environment variable}

\item{value}{The value of the variable}
}
\value{
EnvVariable object
}
\description{
EnvVariable Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An environment variable to set for an action.
}
