% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{ExistingDisk}
\alias{ExistingDisk}
\title{ExistingDisk Object}
\usage{
ExistingDisk(attachment = NULL, source = NULL)
}
\arguments{
\item{attachment}{How the disk will be attached to the Replica}

\item{source}{The name of the Persistent Disk resource}
}
\value{
ExistingDisk object
}
\description{
ExistingDisk Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A pre-existing persistent disk that will be attached to every Replica in the Pool in READ_ONLY mode.
}

