% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{HealthCheck}
\alias{HealthCheck}
\title{HealthCheck Object}
\usage{
HealthCheck(checkIntervalSec = NULL, description = NULL,
  healthyThreshold = NULL, host = NULL, name = NULL, path = NULL,
  port = NULL, timeoutSec = NULL, unhealthyThreshold = NULL)
}
\arguments{
\item{checkIntervalSec}{How often (in seconds) to make HTTP requests for this healthcheck}

\item{description}{The description for this health check}

\item{healthyThreshold}{The number of consecutive health check requests that need to succeed before the replica is considered healthy again}

\item{host}{The value of the host header in the HTTP health check request}

\item{name}{The name of this health check}

\item{path}{The localhost request path to send this health check, in the format /path/to/use}

\item{port}{The TCP port for the health check requests}

\item{timeoutSec}{How long (in seconds) to wait before a timeout failure for this healthcheck}

\item{unhealthyThreshold}{The number of consecutive health check requests that need to fail in order to consider the replica unhealthy}
}
\value{
HealthCheck object
}
\description{
HealthCheck Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

