% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{Label}
\alias{Label}
\title{Label Object}
\usage{
Label(key = NULL, value = NULL)
}
\arguments{
\item{key}{The key for this label}

\item{value}{The value of this label}
}
\value{
Label object
}
\description{
Label Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A label to apply to this replica pool.
}
