% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{Metadata}
\alias{Metadata}
\title{Metadata Object}
\usage{
Metadata(fingerPrint = NULL, items = NULL)
}
\arguments{
\item{fingerPrint}{The fingerprint of the metadata}

\item{items}{A list of metadata items}
}
\value{
Metadata object
}
\description{
Metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Compute Engine metadata entry. Identical to the metadata on the corresponding Compute Engine resource.
}
