% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{MetadataItem}
\alias{MetadataItem}
\title{MetadataItem Object}
\usage{
MetadataItem(key = NULL, value = NULL)
}
\arguments{
\item{key}{A metadata key}

\item{value}{A metadata value}
}
\value{
MetadataItem object
}
\description{
MetadataItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Compute Engine metadata item, defined as a key:value pair. Identical to the metadata on the corresponding Compute Engine resource.
}

