% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{NetworkInterface}
\alias{NetworkInterface}
\title{NetworkInterface Object}
\usage{
NetworkInterface(accessConfigs = NULL, network = NULL, networkIp = NULL)
}
\arguments{
\item{accessConfigs}{An array of configurations for this interface}

\item{network}{Name the Network resource to which this interface applies}

\item{networkIp}{An optional IPV4 internal network address to assign to the instance for this network interface}
}
\value{
NetworkInterface object
}
\description{
NetworkInterface Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Compute Engine NetworkInterface resource. Identical to the NetworkInterface on the corresponding Compute Engine resource.
}
