% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{NewDisk}
\alias{NewDisk}
\title{NewDisk Object}
\usage{
NewDisk(attachment = NULL, autoDelete = NULL, boot = NULL,
  initializeParams = NULL)
}
\arguments{
\item{attachment}{How the disk will be attached to the Replica}

\item{autoDelete}{If true, then this disk will be deleted when the instance is deleted}

\item{boot}{If true, indicates that this is the root persistent disk}

\item{initializeParams}{Create the new disk using these parameters}
}
\value{
NewDisk object
}
\description{
NewDisk Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Persistent Disk resource that will be created and attached to each Replica in the Pool. Each Replica will have a unique persistent disk that is created and attached to that Replica in READ_WRITE mode.
}
