% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{NewDiskInitializeParams}
\alias{NewDiskInitializeParams}
\title{NewDiskInitializeParams Object}
\usage{
NewDiskInitializeParams(diskSizeGb = NULL, diskType = NULL,
  sourceImage = NULL)
}
\arguments{
\item{diskSizeGb}{The size of the created disk in gigabytes}

\item{diskType}{Name of the disk type resource describing which disk type to use to create the disk}

\item{sourceImage}{The name or fully-qualified URL of a source image to use to create this disk}
}
\value{
NewDiskInitializeParams object
}
\description{
NewDiskInitializeParams Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Initialization parameters for creating a new disk.
}
