% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{Pool}
\alias{Pool}
\title{Pool Object}
\usage{
Pool(autoRestart = NULL, baseInstanceName = NULL,
  currentNumReplicas = NULL, description = NULL, healthChecks = NULL,
  initialNumReplicas = NULL, labels = NULL, name = NULL,
  numReplicas = NULL, resourceViews = NULL, selfLink = NULL,
  targetPool = NULL, targetPools = NULL, template = NULL, type = NULL)
}
\arguments{
\item{autoRestart}{Whether replicas in this pool should be restarted if they experience a failure}

\item{baseInstanceName}{The base instance name to use for the replicas in this pool}

\item{currentNumReplicas}{[Output Only] The current number of replicas in the pool}

\item{description}{An optional description of the replica pool}

\item{healthChecks}{Deprecated}

\item{initialNumReplicas}{The initial number of replicas this pool should have}

\item{labels}{A list of labels to attach to this replica pool and all created virtual machines in this replica pool}

\item{name}{The name of the replica pool}

\item{numReplicas}{Deprecated! Use initial_num_replicas instead}

\item{resourceViews}{The list of resource views that should be updated with all the replicas that are managed by this pool}

\item{selfLink}{[Output Only] A self-link to the replica pool}

\item{targetPool}{Deprecated, please use target_pools instead}

\item{targetPools}{A list of target pools to update with the replicas that are managed by this pool}

\item{template}{The template to use when creating replicas in this pool}

\item{type}{Deprecated! Do not set}
}
\value{
Pool object
}
\description{
Pool Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Pool functions: \code{\link{pools.insert}}
}

