% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_functions.R
\name{pools.insert}
\alias{pools.insert}
\title{Inserts a new replica pool.}
\usage{
pools.insert(Pool, projectName, zone)
}
\arguments{
\item{Pool}{The \link{Pool} object to pass to this method}

\item{projectName}{The project ID for this replica pool}

\item{zone}{The zone for this replica pool}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/ndev.cloudman
\item https://www.googleapis.com/auth/replicapool
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/ndev.cloudman, https://www.googleapis.com/auth/replicapool)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/replica-pool/}{Google Documentation}

Other Pool functions: \code{\link{Pool}}
}
