% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{PoolsDeleteRequest}
\alias{PoolsDeleteRequest}
\title{PoolsDeleteRequest Object}
\usage{
PoolsDeleteRequest(abandonInstances = NULL)
}
\arguments{
\item{abandonInstances}{If there are instances you would like to keep, you can specify them here}
}
\value{
PoolsDeleteRequest object
}
\description{
PoolsDeleteRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other PoolsDeleteRequest functions: \code{\link{pools.delete}}
}

