% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{PoolsListResponse}
\alias{PoolsListResponse}
\title{PoolsListResponse Object}
\usage{
PoolsListResponse(nextPageToken = NULL, resources = NULL)
}
\arguments{
\item{nextPageToken}{No description}

\item{resources}{No description}
}
\value{
PoolsListResponse object
}
\description{
PoolsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
