% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{Replica}
\alias{Replica}
\title{Replica Object}
\usage{
Replica(name = NULL, selfLink = NULL, status = NULL)
}
\arguments{
\item{name}{[Output Only] The name of the Replica object}

\item{selfLink}{[Output Only] The self-link of the Replica}

\item{status}{[Output Only] Last known status of the Replica}
}
\value{
Replica object
}
\description{
Replica Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An individual Replica within a Pool. Replicas are automatically created by the replica pool, using the template provided by the user. You cannot directly create replicas.
}
