% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_functions.R
\name{replicas.restart}
\alias{replicas.restart}
\title{Restarts a replica in a pool.}
\usage{
replicas.restart(projectName, zone, poolName, replicaName)
}
\arguments{
\item{projectName}{The project ID for this request}

\item{zone}{The zone where the replica lives}

\item{poolName}{The replica pool name for this request}

\item{replicaName}{The name of the replica for this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/ndev.cloudman
\item https://www.googleapis.com/auth/replicapool
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/ndev.cloudman, https://www.googleapis.com/auth/replicapool)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/replica-pool/}{Google Documentation}
}

