% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{ReplicasDeleteRequest}
\alias{ReplicasDeleteRequest}
\title{ReplicasDeleteRequest Object}
\usage{
ReplicasDeleteRequest(abandonInstance = NULL)
}
\arguments{
\item{abandonInstance}{Whether the instance resource represented by this replica should be deleted or abandoned}
}
\value{
ReplicasDeleteRequest object
}
\description{
ReplicasDeleteRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ReplicasDeleteRequest functions: \code{\link{replicas.delete}}
}
