% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{ReplicasListResponse}
\alias{ReplicasListResponse}
\title{ReplicasListResponse Object}
\usage{
ReplicasListResponse(nextPageToken = NULL, resources = NULL)
}
\arguments{
\item{nextPageToken}{No description}

\item{resources}{No description}
}
\value{
ReplicasListResponse object
}
\description{
ReplicasListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
