% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{ReplicaStatus}
\alias{ReplicaStatus}
\title{ReplicaStatus Object}
\usage{
ReplicaStatus(details = NULL, state = NULL, templateVersion = NULL,
  vmLink = NULL, vmStartTime = NULL)
}
\arguments{
\item{details}{[Output Only] Human-readable details about the current state of the replica}

\item{state}{[Output Only] The state of the Replica}

\item{templateVersion}{[Output Only] The template used to build the replica}

\item{vmLink}{[Output Only] Link to the virtual machine that this Replica represents}

\item{vmStartTime}{[Output Only] The time that this Replica got to the RUNNING state, in RFC 3339 format}
}
\value{
ReplicaStatus object
}
\description{
ReplicaStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The current status of a Replica.
}
