% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{ServiceAccount}
\alias{ServiceAccount}
\title{ServiceAccount Object}
\usage{
ServiceAccount(email = NULL, scopes = NULL)
}
\arguments{
\item{email}{The service account email address, for example: 123845678986@project}

\item{scopes}{The list of OAuth2 scopes to obtain for the service account, for example: https://www}
}
\value{
ServiceAccount object
}
\description{
ServiceAccount Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Compute Engine service account, identical to the Compute Engine resource.
}

