% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{Tag}
\alias{Tag}
\title{Tag Object}
\usage{
Tag(fingerPrint = NULL, items = NULL)
}
\arguments{
\item{fingerPrint}{The fingerprint of the tag}

\item{items}{Items contained in this tag}
}
\value{
Tag object
}
\description{
Tag Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Compute Engine Instance tag, identical to the tags on the corresponding Compute Engine Instance resource.
}

