% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{Template}
\alias{Template}
\title{Template Object}
\usage{
Template(action = NULL, healthChecks = NULL, version = NULL,
  vmParams = NULL)
}
\arguments{
\item{action}{An action to run during initialization of your replicas}

\item{healthChecks}{A list of HTTP Health Checks to configure for this replica pool and all virtual machines in this replica pool}

\item{version}{A free-form string describing the version of this template}

\item{vmParams}{The virtual machine parameters to use for creating replicas}
}
\value{
Template object
}
\description{
Template Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The template used for creating replicas in the pool.
}
\seealso{
Other Template functions: \code{\link{pools.updatetemplate}}
}

