% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicapool_objects.R
\name{VmParams}
\alias{VmParams}
\title{VmParams Object}
\usage{
VmParams(baseInstanceName = NULL, canIpForward = NULL, description = NULL,
  disksToAttach = NULL, disksToCreate = NULL, machineType = NULL,
  metadata = NULL, networkInterfaces = NULL, onHostMaintenance = NULL,
  serviceAccounts = NULL, tags = NULL)
}
\arguments{
\item{baseInstanceName}{Deprecated}

\item{canIpForward}{Enables IP Forwarding, which allows this instance to receive packets destined for a different IP address, and send packets with a different source IP}

\item{description}{An optional textual description of the instance}

\item{disksToAttach}{A list of existing Persistent Disk resources to attach to each replica in the pool}

\item{disksToCreate}{A list of Disk resources to create and attach to each Replica in the Pool}

\item{machineType}{The machine type for this instance}

\item{metadata}{The metadata key/value pairs assigned to this instance}

\item{networkInterfaces}{A list of network interfaces for the instance}

\item{onHostMaintenance}{No description}

\item{serviceAccounts}{A list of Service Accounts to enable for this instance}

\item{tags}{A list of tags to apply to the Google Compute Engine instance to identify resources}
}
\value{
VmParams object
}
\description{
VmParams Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Parameters for creating a Compute Engine Instance resource. Most fields are identical to the corresponding Compute Engine resource.
}
